CLDCONNECT-2407: Refactored according to Devkit v3.7.0 migration guidelines.
Note: Added templates for release-notes.adoc and user-manual.adoc for the time being in-order for project structure validation to pass.
