+++ doc/user-manual.adoc
@@ -1,0 +1,150 @@
+= Anypoint™ Connectors User Guide Template
+
+The following is a proposed connector user guide template for documenting MuleSoft and 3rd party connectors. When completed, a tech writer will work with the development teams to bring existing and new connector documentation to this standard. Please review and add your comments. Thanks!  
+
+Note: Horizontal lines show sections, but won’t be in the actual documents.
+
+---
+
+== Connector Name
+
+The chosen name and one sentence for its description.
+
+---
+
+== Contents:
+
+. Introduction
+.. Prerequisites
+.. Requirements
+.. Dependencies
+... Versions
+... Consumers
+
+. Installing & Configuring
+.. Installing
+.. Configuring
+.. Upgrading from a Previous Version
+
+. Using the Connector
+.. Use Cases and Demos
+.. Best Practices
+.. Tips
+.. Integration Patterns
+.. Adding to a Flow
+.. Query Types
+.. Increasing Performance
+
+. Example Use Case
+..  Anypoint™ Studio  Visual Editor | XML Editor
+.. Example Code
+. Demo
+. F.A.Q
+. See Also
+
+---
+
+== Introduction 
+
+Description and concepts.
+Overview of the system.
+High level description how the connector is providing the connectivity. SOAP/REST/Proprietary protocol. ?
+Description of the main concepts of the system. 
+Integration with other MS capabilities
+
+=== Prerequisites
+
+This document assumes that you are familiar with … +
+To use the this connector, … +
+
+=== Requirements
+
+Hardware, software, and technical requirements. 
+
+=== Dependencies
+
+What MuleSoft software is required to get this connector to work? +
+
+*Versions* +
+
+What versions of JAR files and MS software is required  ?. What is the version sandbox version during the testing ? +
+
+*Consumers* +
+What software is required to user this connector? 
+
+=== Compatibility Matrix
+What system versions is supported ?. Patches/APIs versions
+
+---
+
+== Installing and Configuring 
+
+=== Installing
+
+How do you install this connector? 
+
+=== Configuring
+
+What do you set on configuration? +
+
+=== Updating From an Older Version
+
+What needs to be configured when updating from a previous version of this connector?  +
+Are there configuration items for previous MuleSoft software for this connector?
+ 
+---
+
+== Using This Connector
+
+Outbound, inbound, streaming outbound  +
+
+=== Use Cases and Demos
+List use cases for this connector, including a demo for every use case.+
+
+
+=== Best Practices
+To take full advantage of the functionality … +
+
+=== Tips
+* <Information from Support>
+ 
+=== Integration Patterns
+
+___ recognizes these integration patterns for connecting with other systems… 
+
+=== Adding to a Flow
+
+Use a ____ Connector in your application -> flowchart +
+
+=== Query Types
+
+If talking to a database, what queries are allowed by this connector? +
+
+=== Increasing Performance
+
+What can you set in the connector to increase performance?
+
+---
+
+== Example Use Case
+
+Explain what this example is for. 
+
+It should contain a working fully functionaly Mule application as a demo, where the use case is considered. The demo needs to be simple and only considering a single use case.
+
+===  Anypoint Studio  Visual Editor | XML Editor
+
+Explain how to create this case with  Anypoint Studio  visual editor and XML editor +
+
+=== Code Example
+
+Number code example lines and explain blocks of lines
+ 
+---
+=== F.A.Q
+
+ * Add here a all the items you consider useful.
+
+=== See Also
+
+ * Access full reference documentation for the connector.
