+++ src/main/java/org/mule/module/cmis/model/CMISConnectionType.java
@@ -3,7 +3,7 @@
  * a copy of which has been included with this distribution in the LICENSE.md file.
  */
 
-package org.mule.module.cmis;
+package org.mule.module.cmis.model;
 
 /**
  * Connection Type for CMIS
